Fix bugs in EntityMovement
Fix the Destroy method not working properly because it depended on distance calculation of entity & core and these two were not on the same layer (ergo distance was always larger than 0).

Note to self: Do not mess with Z position in 2D when you want objects to interact with one another properly. Use Sorting Layer for layer arrangement.