Update CoreController
Fix the LinearGenerateShape method. Instead of trying to calculate the radius, it now directly uses the localScale for generateVector.

Note to self: Do not try to complicate matters, for one it's okay to just use the Transform properties even if it's a circle sprite, there's no need to try to calculate the radius of said circle just for the GenerateShape method.