Update LevelTick.cs
Now the level tick will update in runtime according to currentLevel.
Note that only the upward movement is implemented because the assumption is there will be no regress back to lower levels. Player loses the game once they've run out of vitalityPoint, regardless of which level they were currently at.